refactor: 투수 이닝 데이터 타입 변경 및 야구 표기법 변환 로직 추가
- `PitcherRecordDto`의 `innings` 필드 타입을 `Int`에서 `Double`로 변경
- `Double.toInningString()` 확장 함수 추가를 통해 소수점 형태의 이닝을 야구 전문 표기법(⅓, ⅔)으로 변환하는 로직 구현
- `CheckInUiMapper`에서 투수 레코드 변환 시 `toInningString()`을 적용하여 UI 표시 로직 개선

+++ mobile/composeApp/src/commonMain/kotlin/com/yagubogu/data/dto/response/checkin/PitcherRecordDto.kt
@@ -10,7 +10,7 @@ data class PitcherRecordDto(
     @SerialName("result")
     val result: String, // 등판 결과
     @SerialName("innings")
-    val innings: Int, // 투구 이닝 (예: 6.2 = 6 2/3이닝)
+    val innings: Double, // 투구 이닝 (예: 6.2 = 6 2/3이닝)
     @SerialName("pitchCount")
     val pitchCount: Int, // 총 투구 수
     @SerialName("hitsAllowed")
